% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_avg_sensitivities}
\alias{compute_avg_sensitivities}
\title{Compute average sensitivity of each compound}
\usage{
compute_avg_sensitivities(data_list, sens_column = "auc_avg")
}
\arguments{
\item{sens_column}{}
}
\description{
Compute average sensitivity of each compound
}
\examples{
compute_avg_sensitivities()
}
