% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_sensitivity_selectivities}
\alias{compute_sensitivity_selectivities}
\title{Compute selectivity of the killing of each compound}
\usage{
compute_sensitivity_selectivities(data_list, sens_column = "auc_avg")
}
\arguments{
\item{sens_column}{}
}
\description{
Compute selectivity of the killing of each compound
}
\examples{
compute_sensitivity_selectivities()
}
