% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_tas_viability_cors}
\alias{compute_tas_viability_cors}
\title{Compute correlations between TAS and sensitivity}
\usage{
compute_tas_viability_cors(data_list, sens_column = "auc_avg")
}
\arguments{
\item{sens_column}{}
}
\description{
Compute correlations between TAS and sensitivity
}
\examples{
compute_tas_viability_cors()
}
