% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_viability_signature}
\alias{compute_viability_signature}
\title{Compute viability signature}
\usage{
compute_viability_signature(data_list, compound = NA, global = F,
  sens_column = "auc_avg")
}
\arguments{
\item{sens_column}{}
}
\description{
Compute viability signature
}
\examples{
compute_viability_signature()
}
