% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_viability_signature_all_cpds}
\alias{compute_viability_signature_all_cpds}
\title{Compute a unique viability signature for each drug}
\usage{
compute_viability_signature_all_cpds(data_list, sens_column = "auc_avg",
  selectivity_threshold = 0, covariates = "")
}
\arguments{
\item{sens_column}{}
}
\description{
Compute a unique viability signature for each drug
}
\examples{
compute_viability_signature_all_cpds()
}
