% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_viability_signature_pearson_all_cpds}
\alias{compute_viability_signature_pearson_all_cpds}
\title{Compute a unique viability signature for each drug using Pearson correlation (other function uses limma)}
\usage{
compute_viability_signature_pearson_all_cpds(data_list,
  sens_column = "auc_avg")
}
\arguments{
\item{sens_column}{}
}
\description{
Compute a unique viability signature for each drug using Pearson correlation (other function uses limma)
}
\examples{
compute_viability_signature_pearson_all_cpds()
}
