% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{compute_viability_signature_strengths}
\alias{compute_viability_signature_strengths}
\title{Compute the strength of each compound's viability signature}
\usage{
compute_viability_signature_strengths(via_sig_list,
  which_sig = "viability_related")
}
\arguments{
\item{which_sig}{}
}
\description{
Compute the strength of each compound's viability signature
}
\examples{
compute_viability_signature_strengths()
}
