% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{get_cpd_viability_signature_matrix}
\alias{get_cpd_viability_signature_matrix}
\title{Get a matrix of viability sigantures given a list of them}
\usage{
get_cpd_viability_signature_matrix(via_sig_list,
  which_sig = "viability_related")
}
\arguments{
\item{which_sig}{}
}
\description{
Get a matrix of viability sigantures given a list of them
}
\examples{
get_cpd_viability_signature_matrix()
}
