% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_gene_set_genes}
\alias{get_gene_set_genes}
\title{Fetch the gene names belonging to a certain gene set}
\usage{
get_gene_set_genes(gsc, collection_name, gene_set_name)
}
\description{
Fetch the gene names belonging to a certain gene set
}
\examples{
get_gene_set_genes()
}
