% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_l1000_cpd_data}
\alias{load_l1000_cpd_data}
\title{Load L1000 compound data from Taiga (includes matched viability data)}
\usage{
load_l1000_cpd_data(normalize_sensitivity = TRUE)
}
\description{
Load L1000 compound data from Taiga (includes matched viability data)
}
\examples{
load_l1000_cpd_data()
}
