% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_l1000_shrna_data}
\alias{load_l1000_shrna_data}
\title{Load L1000 shRNA data from Taiga (includes matched viability data)}
\usage{
load_l1000_shrna_data()
}
\description{
Load L1000 shRNA data from Taiga (includes matched viability data)
}
\examples{
load_l1000_shrna_data()
}
