% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_repurposing_cpd_annotations}
\alias{load_repurposing_cpd_annotations}
\title{Load compound annotations (MOA, target, etc.)}
\usage{
load_repurposing_cpd_annotations()
}
\description{
Load compound annotations (MOA, target, etc.)
}
\examples{
load_repurposing_cpd_annotations()
}
