% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_via_sig_volcano}
\alias{make_via_sig_volcano}
\title{Plot viability signature output}
\usage{
make_via_sig_volcano(via_sig_output, cpd_name)
}
\description{
Plot viability signature output
}
\examples{
make_via_sig_volcano()
}
