% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cpd_and_shrna_sigs}
\alias{plot_cpd_and_shrna_sigs}
\title{Plot comparison of one cpd}
\usage{
plot_cpd_and_shrna_sigs(cpd_via_sig_df, shrna_via_sig_df, cpd_name,
  shrna_name, shared_pathway_genes = NULL)
}
\description{
Plot comparison of one cpd
}
\examples{
plot_cpd_and_shrna_sigs()
}
