% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_auc_relationship}
\alias{plot_gene_auc_relationship}
\title{Plot an expression-sensitivity relationship for one gene}
\usage{
plot_gene_auc_relationship(cpd, via_sig, data_list, gene,
  sens_column = "auc_avg", is_plot_abline = T,
  is_plot_intercept_dot = T)
}
\description{
Plot an expression-sensitivity relationship for one gene
}
\examples{
plot_gene_auc_relationship()
}
