% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_two_via_sigs}
\alias{plot_two_via_sigs}
\title{Plot comparison of two compounds' viability signatures}
\usage{
plot_two_via_sigs(all_via_sig_df, cpd1, cpd2)
}
\description{
Plot comparison of two compounds' viability signatures
}
\examples{
plot_two_via_sigs()
}
