% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_via_sig_comparison}
\alias{plot_via_sig_comparison}
\title{Plot viability-related and -unrelated signatures against each other}
\usage{
plot_via_sig_comparison(via_sig_output, gene_class_df = NULL,
  cpd_to_plot = NULL)
}
\description{
Plot viability-related and -unrelated signatures against each other
}
\examples{
plot_via_sig_comparison()
}
