% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{quantile_normalize_viability_data}
\alias{quantile_normalize_viability_data}
\title{Quantile normalize a dataset so that each column has the same distribution (typically used for normalizing across sensitivity datasets)}
\usage{
quantile_normalize_viability_data(data_df)
}
\arguments{
\item{data_df}{dataframe to be normalized}
}
\description{
Quantile normalize a dataset so that each column has the same distribution (typically used for normalizing across sensitivity datasets)
}
\examples{
quantile_normalize_viability_data()
}
