% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viability_signature_modeling.R
\name{run_limma_via_sig}
\alias{run_limma_via_sig}
\title{Run limma linear model to compute viability signature}
\usage{
run_limma_via_sig(data_df, gene_names, sens_column = "auc_avg",
  covariates = "")
}
\arguments{
\item{sens_column}{}
}
\description{
Run limma linear model to compute viability signature
}
\examples{
run_limma_via_sig()
}
