% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_publication_plot}
\alias{save_publication_plot}
\title{Save plots in a standardized way}
\usage{
save_publication_plot(ggplot_obj, filename)
}
\description{
Save plots in a standardized way
}
\examples{
save_publication_plot()
}
