% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_scatterplot}
\alias{theme_scatterplot}
\title{Change aesthetics of scatterplot to a standard style}
\usage{
theme_scatterplot()
}
\description{
Change aesthetics of scatterplot to a standard style
}
\examples{
theme_scatterplot()
}
